Moved all actions out of FeaturesContainer, TasksContainer to individual task
