+++ client/components/sections/EstimationTask.js
@@ -4,6 +4,12 @@ import PropTypes from 'prop-types'
 import * as logger from '../../clientLogger'
 import * as SC from '../../../server/serverconstants'
 import _ from 'lodash'
+import * as A from '../../actions'
+import * as COC from '../../components/componentConsts'
+import {connect} from 'react-redux'
+import {initialize} from 'redux-form'
+import {NotificationManager} from 'react-notifications'
+import * as EC from "../../../server/errorcodes";
 
 class EstimationTask extends React.PureComponent {
 
@@ -72,8 +78,7 @@ class EstimationTask extends React.PureComponent {
                                              onClick={() => {
                                                  this.props.deleteTask(task)
                                              }}/> :
-                    <img key="he_requested_delete" src="/images/he_requested_delete_disable.png"
-                    />)
+                    <img key="he_requested_delete" src="/images/he_requested_delete_disable.png"/>)
             } else {
                 // Negotiator can delete any task during its review without getting permission from estimator
                 buttons.push(editView ? <img key="delete" src="/images/delete.png"
@@ -82,8 +87,8 @@ class EstimationTask extends React.PureComponent {
                                              }}/> : <img key="delete" src="/images/delete_disable.png"/>)
             }
 
-        } else if (loggedInUserRole == SC.ROLE_ESTIMATOR) {
-            if (task.addedInThisIteration && task.owner == SC.OWNER_ESTIMATOR) {
+        } else if (loggedInUserRole === SC.ROLE_ESTIMATOR) {
+            if (task.addedInThisIteration && task.owner === SC.OWNER_ESTIMATOR) {
                 // As estimator has added this task in this iteration only, he/she would be able to edit/delete it without any restrictions
                 buttons.push(editView ? <img key="edit" src="/images/edit.png"
                                              onClick={() => {
@@ -161,8 +166,8 @@ class EstimationTask extends React.PureComponent {
             }
         }
 
-        if (loggedInUserRole == SC.ROLE_NEGOTIATOR && _.includes([SC.STATUS_INITIATED, SC.STATUS_REVIEW_REQUESTED], estimationStatus) ||
-            loggedInUserRole == SC.ROLE_ESTIMATOR && _.includes([SC.STATUS_ESTIMATION_REQUESTED, SC.STATUS_CHANGE_REQUESTED], estimationStatus)) {
+        if (loggedInUserRole === SC.ROLE_NEGOTIATOR && _.includes([SC.STATUS_INITIATED, SC.STATUS_REVIEW_REQUESTED], estimationStatus) ||
+            loggedInUserRole === SC.ROLE_ESTIMATOR && _.includes([SC.STATUS_ESTIMATION_REQUESTED, SC.STATUS_CHANGE_REQUESTED], estimationStatus)) {
             if (task.feature && task.feature._id) {
                 // This task is part of some feature so add move out of feature button
                 buttons
@@ -192,25 +197,25 @@ class EstimationTask extends React.PureComponent {
             </div>
             <div className="col-md-3">
                 {task.owner == SC.OWNER_ESTIMATOR && task.addedInThisIteration && <div className="flagStrip">
-                    <img src="/images/estimator_new_flag.png" title="Added by Estimator"></img>
+                    <img src="/images/estimator_new_flag.png" title="Added by Estimator"/>
                 </div>}
 
                 {task.owner == SC.OWNER_NEGOTIATOR && task.addedInThisIteration && <div className="flagStrip">
-                    <img src="/images/negotiator_new_flag.png" title="Added by Negotiator"></img>
+                    <img src="/images/negotiator_new_flag.png" title="Added by Negotiator"/>
                 </div>}
 
                 {!task.repo.addedFromThisEstimation &&
                 <div className="flagStrip">
-                    <img src="/images/repo_flag.png" title="From Repository"></img>
+                    <img src="/images/repo_flag.png" title="From Repository"/>
                 </div>
                 }
 
                 {task.estimator.changedInThisIteration && <div className="flagStrip">
-                    <img src="/images/estimator_edit_flag.png" title="Edited by Estimator"></img>
+                    <img src="/images/estimator_edit_flag.png" title="Edited by Estimator"/>
                 </div>}
 
                 {task.negotiator.changedInThisIteration && <div className="flagStrip">
-                    <img src="/images/negotiator_edit_flag.png" title="Edited by Negotiator"></img>
+                    <img src="/images/negotiator_edit_flag.png" title="Edited by Negotiator"/>
                 </div>}
 
 
@@ -248,5 +253,124 @@ EstimationTask.defaultProps = {
     expanded: false,
     isFeatureTask: false
 }
+
+EstimationTask = connect(null, (dispatch, ownProps) => ({
+    toggleEditRequest: (values) => {
+        let task = {}
+        task.task_id = values._id
+        return dispatch(A.requestForTaskEditPermissionOnServer(task)).then(json => {
+            if (json.success) {
+
+                if (json.data && json.data.estimator && json.data.estimator.changeRequested)
+                    NotificationManager.success("Edit request on Task raised...")
+                else
+                    NotificationManager.success("Edit request on Task cleared...")
+            } else {
+                NotificationManager.error("Unknown error occurred")
+            }
+        })
+    },
+    deleteTask: (values) => {
+        return dispatch(A.deleteEstimationTaskOnServer(values.estimation._id, values._id)).then(json => {
+            if (json.success) {
+                NotificationManager.success("Task Deleted successfully")
+            }
+            else
+                NotificationManager.error("Task Deletion Failed !")
+
+        })
+    },
+    toggleDeleteRequest: (values) => {
+        let task = {}
+        task.task_id = values._id
+        return dispatch(A.requestForTaskDeletePermissionOnServer(task)).then(json => {
+            if (json.success) {
+                NotificationManager.success("Task Delete requested successfully")
+            } else {
+                if (json.code == EC.INVALID_OPERATION)
+                    NotificationManager.error("Task Delete already requested")
+                else
+                    NotificationManager.error("Unknown error occurred")
+            }
+        })
+    },
+    editTask: (values, loggedInUserRole) => {
+        dispatch(A.showComponent(COC.ESTIMATION_TASK_DIALOG))
+        let task = {}
+        task._id = values._id
+        task.estimation = values.estimation
+        if (loggedInUserRole != SC.ROLE_NEGOTIATOR) {
+            task.name = values.estimator.name
+            task.description = values.estimator.description
+            task.estimatedHours = values.estimator.estimatedHours
+        }
+        else {
+            task.name = values.negotiator.name
+            task.description = values.negotiator.description
+            task.estimatedHours = values.negotiator.estimatedHours
+        }
+        task.feature = values.feature
+        task.technologies = values.technologies
+        dispatch(initialize("estimation-task", task))
+    },
+    moveToFeature: (values) => {
+        let task = {}
+        task.task_id = values._id
+        dispatch(A.showComponent(COC.MOVE_TASK_TO_FEATURE_FORM_DIALOG))
+        dispatch(initialize("MoveTaskInFeatureForm", task))
+    },
+    moveTaskOutOfFeature: (values) => {
+        let task = {}
+        task.task_id = values._id
+        task.feature_id = values.feature._id
+        dispatch(A.moveTaskOutOfFeatureOnServer(values)).then(json => {
+            if (json.success) {
+                NotificationManager.success('Task moved out of feature Successfully')
+            } else {
+                NotificationManager.success('Some error occurred')
+            }
+
+        })
+        //dispatch(initialize("MoveTaskInFeatureForm", task))
+    },
+    toggleGrantEdit: (values) => {
+        let task = {}
+        task.task_id = values._id
+        return dispatch(A.grantEditPermissionOfTaskOnServer(task))
+    },
+    openTaskSuggestionForm: (values, loggedInUserRole) => {
+        let task = {
+            loggedInUserRole: loggedInUserRole,
+            readOnly: {}
+        }
+        task._id = values._id
+        task.estimation = values.estimation
+        if (loggedInUserRole == SC.ROLE_NEGOTIATOR) {
+            /* Since negotiator is logged in, he would see details of negotiator section in editable form and details of estimator section in read only form  */
+            task.name = values.negotiator.name
+            task.description = values.negotiator.description
+            task.estimatedHours = values.negotiator.estimatedHours
+
+            task.readOnly.name = values.estimator.name
+            task.readOnly.description = values.estimator.description
+            task.readOnly.estimatedHours = values.estimator.estimatedHours
+
+        } else if (loggedInUserRole == SC.ROLE_ESTIMATOR) {
+            /* Since estimator is logged in, he would see details of  estimator section in editable form  */
+            task.name = values.estimator.name
+            task.description = values.estimator.description
+            task.estimatedHours = values.estimator.estimatedHours
+
+            task.readOnly.name = values.negotiator.name
+            task.readOnly.description = values.negotiator.description
+            task.readOnly.estimatedHours = values.negotiator.estimatedHours
+
+        }
+
+        dispatch(initialize("estimation-suggest-task", task))
+        dispatch(A.showComponent(COC.ESTIMATION_SUGGEST_TASK_FORM_DIALOG))
+    }
+}))(EstimationTask)
+
 
 export default EstimationTask
